requirements.txt
Flask
PyJWT
bcrypt
python-dotenv